Fix a bug in staleness detection.
When an empty lock file was left, it wasn't considered as stale. And
thus applications could wait on unowned locks forever.

This was particularly bad for upgrading CCM. Old versions don't use
CAF for locking, but the newest ones do. When doing this upgrade, CCM
would just break on the target machine, locking forever.